Fix substitution of aliases with leading space
A command like 'help s', which should've been interpreted as 'help step'
(i.e., access the documentation for the step command) was instead being
interpreted as the nonsensical 'helpstep' since the regex substitution
used to get from the alias 's' to 'step' was swallowing up the leading
whitespace character. Changing the substitution text to include a
leading space fixes this for the two aliases affected: 's' ('step') and
'sm' ('stepm').

The third available alias, which makes '123' equivalent to 'stepm 123',
now works with leading whitespace.