Fix confirm. to delete all marks when none defined
When no marks are defined, the command `delete marks` should yield the
response "No marks to delete." However, if marks had been defined and
subsequently deleted, `delete marks` instead used its behavior for when
some marks are defined (that is, it asks for confirmation to delete them
all). This happened because `delete marks` was just checking if the dict
holding the marks was empty, and deleting individual marks never
actually removed the dict entry for a given key---it just cleared or
subtracted from the set of candidates associated with that key.