Make reasons return set; clarify documentation
The `reasons` method in Solver now returns a set instead of a list to
underline that the order of the iterable's elements is unimportant. No
code that touched the return value of `reasons` ever had a reason for it
to be a list, so the only changes related to that code were to the
documentation to modify requirements from "list" to merely "iterable".
Also, documentation of methods in Solver that return lists instead of
sets was altered slightly to indicate that they specifically return
*ordered* elements.