Allow last move type to be overridden in `reasons`
Add a new optional parameter to the `reasons` method in Solver that
allows the type of the last move to be given instead of looked up in
`move_history`. This allows reasons to be generated for a move with a
type usually ignored by the `reasons` method (e.g., `MoveType.MANUAL`)
by telling that method (via the new parameter) how to interpret the
move. To slightly simplify some of the code involved (both in this
commit and the forthcoming commit that allows Controller to take
advantage of this commit), a new attribute was added to Solver,
DEDUCTIVE_MOVE_TYPES, which is a list of move types that `reasons` knows
how to handle.