Make `explain` attempt to explain manual moves
The `explain` command now attempts to find reasons for moves with the
type `MANUAL` (i.e., those made using `stepm`) by temporarily forcing,
in turn, each deductive move type onto the move and seeing which of
those types yields the largest location set from `Solver.reasons()`.
Unfortunately, this forces a strange distinction between the actual type
of the last move and the type needed to make sense of that location set.
To deal with this, the method that prints the puzzle with the reasons
colored and its backend method that determines which locations need
to be mapped to which colors both had to add new parameters and, in the
latter case, become a good bit more complex. Before, for example, a
row-wise deduction meant that all blank cells in the same row as the
last move could be colored as nonviable locations since the solver would
not have made the move had that not been true; now, as little as one
blank in that row may be nonviable since the user isn't restricted to
only making necessarily correct moves. Thus, each location in the
location set has to be checked to see what row(s) or column(s) it makes
nonviable (the plurals arise from locations in the same stack or band as
the move).