Add step commands prioritizing row, col, or box
Add the commands `stepr ROW`, `stepc COL`, and `stepb BOX`, which
prioritize finding deductions in the row, column, or box (respectively)
passed to them. In light of the new `BOXWISE` move type in Solver that
makes that last command viable, modify `_get_reasons_colormap` to handle
colorizing that type of move.