Make `candidates` mirror `step`-able deductions
Improve the consistency between which locations `candidates` claims have
only one possible number and those locations for which `step` can actually
deduce that number. To this end, add a new deductive move type
`ELIMINATION`, which allows deductions based on a number being the only
one possible at a location. For example, in the following Sudoku, the
number 3 could be deduced at row 3, column 4 because the numbers in its
box prevent all but {1, 3} as possible numbers, and the 1 in its same
row means only 3 is actually possible:

.........................
! 3 2 6 ! 7 9 4 ! 8 1 5 !
! 4 5 9 ! _ 6 8 ! _ 7 _ !
! 7 1 8 ! _ 5 2 ! _ _ _ !
!.......!.......!.......!
! 8 9 7 ! _ 4 _ ! 5 _ 3 !
! 2 _ _ ! 5 3 9 ! _ 8 _ !
! _ 4 _ ! _ 8 _ ! _ _ _ !
!.......!.......!.......!
! 9 _ _ ! _ _ _ ! _ _ _ !
! 6 7 2 ! 4 1 3 ! 9 5 8 !
! 1 _ _ ! _ 7 6 ! _ _ _ !
!.......!.......!.......!

Thus, make `step` attempt the `ELIMINATION`-style deductions that
`candidates` had always been reporting. Also, make `candidates` first
look for possible numbers in the same necessary-move cache `step` uses,
and only have it fall back on `possible_next_moves`, which was rewritten
with this purpose in mind, when necessary. Finally, add code to handle
the new `ELIMINATION` move type in `reasons`.