Fix `x`/`info mark` behavior if given bad location
If given one or more invalid (row, column) locations, the commands `x`
and `info mark` are supposed to operate as normal on any remaining valid
locations and then report the invalid one(s). Instead, when faced with
this situation, both merely reported an invalid location and exited. To
fix this, an optional parameter was added to the method both call to
parse the locations that tells that method not to bother validating the
locations it finds (though it still defaults to validating them since
other methods rely on that behavior).