Improve documentation consistency (and clarity)
Add a lone newline to the end of each docstring to be consistent with
the intended style of documentation (numpy style). Wrap all comments to
75 characters to be consistent with the (numpy-style-specified)
docstring width. Add any missing module docstrings to be consistent with
PEP 257's recommendation that all modules have docstrings. Make
minor---and, for `cmdmap`, somewhat more major---clarifications to some
documentation.