Fix `report_errors` and `error_count` bugs
In both methods, fix incorrect assumption that `log_entry` raises a
KeyError for objects not in `log`---it actually just returns 0. In
`report_errors`, this mistaken assumption prevented an object with no
associated errors from having "no errors" reported. In `error_count`,
the assumption did not affect behavior. Also, in `report_errors` fix
ANDing `err` instead of the number associated with it, `err.errno`; this
caused a crash when the object passed to this method actually had any
errors associated with it.