Fix crash on `ImportError`
When trying to read a Sudoku from an image, if importing `sudokuimg` or
any of the modules it requires raises an `ImportError`, the module that
caused the error is supposed to be reported and then the program should
continue. Instead, the program crashed with an `AttributeError` because
I was mistakenly using `strerror` in trying to report the error, and
`strerror` is not an attribute of `ImportError`.