Add code to make symmetric puzzles; alter a retval
Add a function that changes a puzzle into one with 180-degree
rotational symmetry by adding and, optionally, removing clues. Add a
parameter to `generate` to allow it to generate these types of puzzles.
Make the return argument of `minimize` negative to better match the
return value in the new symmetry function, which is negative when the
new, symmetric puzzle it creates has fewer clues than the original and
is positive when it has more.