Change default to not using ANSI escape sequences
Now `strfboard` uses no ANSI escape sequences in the string it returns
unless its new optional parameter, `ansi_mode`, is set to `True`.
Previously, it made minimal use of these sequences to make dim the
numbers in the board's axes (when `show_axes=True`) and to make the
clues in the bigger versions of the board bold (when `candidate_map` is
not `None`).