Guarantee consistent retvals if multiple solutions
Define the order of iteration over (by definition, unordered) sets in
`_backtrack`, `_exact_hitting_set`, and related methods so that the
methods that ultimately use them always return or yield the same
value(s) in, for yields, the same order. The assumptions about set
iteration order never seemed to affect anything in Python 2, but in
Python 3 it was causing the generator to return different puzzles for
the same seed due to its reliance on `autosolve_without_history`, one of
the affected methods. Note that this commit breaks some tests of the
generator since it changes what puzzles are associated with what seeds.