Avoid ANSI in error report for inconsistent boards
When attempting to import a Sudoku with inconsistencies, instead of
printing the full board with the inconsistencies highlighted in red
(using ANSI color codes), now print a board with only the inconsistent
clues on it. This allows the error to be viewable in more contexts and
also prevents any peculiar wrapping on the lines making up the board
since those ANSI codes count in `textwrap` as part of the line length.